+++ sumR/R/data_transformation.R
@@ -1,14 +1,4 @@
-#' @title data imputation
-#' @description This function apply data imputation from 1 to selected noise level
-#' @param data dataframe
-#' @param noise numerical value for the noise level
-#' @param seed global seed for reproducible results
-#' @importFrom stats runif
-data_imputation <- function(data, noise, seed) {
-  set.seed(seed)
-  data[data == 0] <- runif(1, min = 1, max = noise)
-  return(data)
-}
+