+++ sumR/R/linear_regression_func.R
@@ -1,3 +1,67 @@
+#' @title Model glmnet
+#' @export
+modelGLM <- function(exp, classifiers = metadata(exp)$phenotype, assay = 1,
+                        type.measure = "deviance", lambda = c("lambda.min", "lambda.1se"),
+                        cv = 5, ratio = 0.8, seed = NULL, alphas = seq(0, 1, 0.1)){
+  if (is.null(classifiers)) stop("Cannot perform test without classifiers")
+  if (!is.null(seed)) set.seed(seed)
+
+  data <- assay(exp, assay)
+  samples <- as.factor(exp[[classifiers]])
+  split <- caTools::sample.split(samples, SplitRatio = ratio)
+  model(exp, "glmnet")$train <- colnames(exp)[which(split)]
+  model(exp, "glmnet")$test <- colnames(exp)[-which(split)]
+
+
+  training <- t(data[, model(exp, "glmnet")$train])
+  test <- t(data[, model(exp, "glmnet")$test])
+
+  folds <- caret::createFolds(y = model(exp, "glmnet")$train,
+                              k = cv, list = F)
+  alpha <- alphas[which.min(vapply(alphas, function(a){
+    tryCatch({
+      return(
+        cv.glmnet(
+          y = exp[, model(exp, "glmnet")$train][[classifiers]],
+          x = as.matrix(scale(training)), foldid = folds,
+          family = "binomial", alpha = a, type.measure = type.measure,
+          nfolds = cv
+        )[[lambda[1]]]
+      )
+    }, error = function(x) return(Inf))
+
+  }, double(1)))]
+  m <- cv.glmnet(
+    y = exp[, model(exp, "glmnet")$train][[classifiers]],
+    x = as.matrix(scale(training)), foldid = folds,
+    family = "binomial", alpha = alpha, type.measure = type.measure,
+    nfolds = cv
+  )
+  model(exp, "glmnet")$model <- m
+
+  model(exp, "glmnet")$predict <- predict(m, s = m[[lambda[1]]],
+                                          newx = as.matrix(scale(test)), type = "class")
+  exp
+  # do.call(rbind, lapply(0:10, function(i){
+  #   m <- cv.glmnet(
+  #     y = exp[, model(exp, "glmnet")$train][[classifiers]],
+  #     x = as.matrix(scale(training)), foldid = folds,
+  #     family = "binomial", alpha = i / 10, type.measure = type.measure,
+  #     nfolds = cv
+  #   )
+  #
+  #
+  #   predict <- predict(m, s = m[[lambda[1]]],
+  #                      newx = as.matrix(scale(test)),
+  #                      type = type.measure
+  #   )
+  #
+  #
+  #   cf <- confusionMatrix(table(data = predict, reference = exp[, model(exp, "glmnet")$test][[classifiers]]))
+  #   data.frame(alpha = i / 10, cf_acc = cf$overall$Accuracy)
+  # }))
+}
+
 ## choosing alpha value for elastic net
 #' @title Alpha value choice
 #' @description This function choose the alpha value for elastic net regression with the desired accuracy
@@ -9,46 +73,23 @@
 #' @importFrom glmnet cv.glmnet
 #' @importFrom caret confusionMatrix
 #' @importFrom dplyr select
-glmnet_cv_alphachoice <- function(training, test, type.measure, seed, lambda = c("lambda.min", "lambda.1se")) {
-  list_fits <- list()
-  for (i in 0:10) {
-    fit_name <- paste0("alpha", i / 10)
-    set.seed(seed)
-    list_fits[[fit_name]] <- cv.glmnet(
+glmnet_cv_alphachoice <- function(training, test, type.measure, seed = NULL, lambda = c("lambda.min", "lambda.1se")) {
+  #list_fits <- list()
+  if (!is.null(seed)) set.seed(seed)
+
+  do.call(rbind, lapply(0:10, function(i){
+    model <- cv.glmnet(
       y = training$samples,
-      x = as.matrix(scale(training %>% dplyr::select(-samples))),
+      x = as.matrix(scale(training)),
       family = "binomial", alpha = i / 10, type.measure = type.measure
     )
-  }
-  if (lambda[1] == "lambda.min") {
-    results <- data.frame()
-    for (i in 0:10) {
-      fit_name <- paste0("alpha", i / 10)
-      predict <- predict(list_fits[[fit_name]],
-        s = list_fits[[fit_name]]$lambda.min, newx = as.matrix(scale(test %>% dplyr::select(-samples))),
-        type = "class"
-      )
-      cf <- confusionMatrix(table(data = predict, reference = test$samples))
-      cf_acc <- cf[["overall"]][["Accuracy"]]
-      temp <- data.frame(alpha = i / 10, cf_acc = cf_acc, fit_name = fit_name)
-      results <- rbind(results, temp)
-    }
-    return(results)
-  } else if (lambda[1] == "lambda.1se") {
-    results <- data.frame()
-    for (i in 0:10) {
-      fit_name <- paste0("alpha", i / 10)
-      predict <- predict(list_fits[[fit_name]],
-        s = list_fits[[fit_name]]$lambda.1se, newx = as.matrix(scale(test %>% dplyr::select(-samples))),
-        type = "class"
-      )
-      cf <- confusionMatrix(table(data = predict, reference = test$samples))
-      cf_acc <- cf[["overall"]][["Accuracy"]]
-      temp <- data.frame(alpha = i / 10, cf_acc = cf_acc, fit_name = fit_name)
-      results <- rbind(results, temp)
-    }
-    return(results)
-  }
+    predict <- predict(model, s = model[[lambda[1]]],
+                       newx = as.matrix(scale(test)),
+                       type = type.measure
+    )
+    cf <- confusionMatrix(table(data = predict, reference = test$samples))
+    data.frame(alpha = i / 10, cf_acc = cf$overall$Accuracy)
+  }))
 }
 
 
@@ -62,15 +103,14 @@ glmnet_cv_alphachoice <- function(training, test, type.measure, seed, lambda = c
 #' @param nfolds number of folds for cross validation
 #' @importFrom glmnet cv.glmnet
 #' @importFrom dplyr select
-final_glmnet <- function(training, test, alpha, type.measure, seed, nfolds) {
-  set.seed(seed)
-  fit_final <- cv.glmnet(
+final_glmnet <- function(training, test, alpha, type.measure, seed = NULL, nfolds = 10) {
+  if (!is.null(seed)) set.seed(seed)
+  cv.glmnet(
     y = training$samples,
     x = as.matrix(scale(training %>% dplyr::select(-samples))),
-    family = "binomial", alpha = alpha, type.measure = type.measure, keep = T, nfolds = nfolds
+    family = "binomial", alpha = alpha, type.measure = type.measure,
+    keep = T, nfolds = nfolds
   )
-
-  return("final_fit" = fit_final)
 }
 
 
@@ -88,7 +128,8 @@ final_glmnet_plot <- function(fit_final) {
 #' @importFrom dplyr select
 #' @importFrom glmnet assess.glmnet
 final_glmnet_assess <- function(fit_final, test, lambda = "lambda.min") {
-  assess.glmnet(fit_final, newx = as.matrix(scale((test %>% dplyr::select(-samples)))), newy = test$samples, s = lambda)
+  assess.glmnet(fit_final, newx = as.matrix(scale((test %>% dplyr::select(-samples)))),
+                newy = test$samples, s = lambda)
 }
 
 
@@ -100,7 +141,9 @@ final_glmnet_assess <- function(fit_final, test, lambda = "lambda.min") {
 #' @importFrom caret confusionMatrix
 #' @importFrom dplyr select
 final_glmnet_confusionmatrix <- function(fit_final, test, lambda = "lambda.min") {
-  predict <- predict(fit_final, s = lambda, newx = as.matrix(scale((test %>% dplyr::select(-samples)))), type = "class")
+  predict <- predict(fit_final, s = lambda,
+                     newx = as.matrix(scale((test %>% dplyr::select(-samples)))),
+                     type = "class")
   confusionMatrix(table(data = predict, reference = test$samples))
 }